code cleanup

+++ applications/tetris_game/tetris_game.c
@@ -328,7 +328,7 @@ static void tetris_game_process_step(TetrisState* tetris_state, Piece* newPiece,
                     }
                 }
 
-                uint32_t oldNumLines = tetris_state->numLines;
+                uint16_t oldNumLines = tetris_state->numLines;
                 tetris_state->numLines += numLines;
                 if((oldNumLines / 10) % 10 != (tetris_state->numLines / 10) % 10) {
                     tetris_state->fallSpeed -= 50;
@@ -460,57 +460,6 @@ int32_t tetris_game_app() {
         }
 
         tetris_game_process_step(tetris_state, newPiece, wasDownMove);
-
-        // tetris_game_remove_curr_piece(tetris_state);
-
-        // if(wasDownMove) {
-        //     if(tetris_game_piece_at_bottom(tetris_state, newPiece)) {
-        //         osTimerStop(timer);
-
-        //         tetris_game_render_curr_piece(tetris_state);
-        //         uint8_t numLines = 0;
-        //         uint8_t lines[] = { 0,0,0,0 };
-
-        //         tetris_game_check_for_lines(tetris_state, lines, &numLines);
-        //         if(numLines > 0) {
-        //             for(int i = 0; i < numLines; i++) {
-
-        //                 // zero out row
-        //                 for(int j = 0; j < FIELD_WIDTH; j++) {
-        //                     tetris_state->playField[lines[i]][j] = false;
-        //                 }
-        //                 // move all above rows down
-        //                 for(int k = lines[i]; k >= 0 ; k--) {
-        //                     for(int m = 0; m < FIELD_WIDTH; m++) {
-        //                         tetris_state->playField[k][m] = (k == 0) ? false : tetris_state->playField[k-1][m];
-        //                     }
-        //                 }
-        //             }
-
-        //             uint32_t oldNumLines = tetris_state->numLines;
-        //             tetris_state->numLines += numLines;
-        //             if((oldNumLines / 10) % 10 != (tetris_state->numLines / 10) % 10) {
-        //                 tetris_state->fallSpeed -= 50;
-        //             }
-        //         }
-
-        //         // Check for game over
-        //         Piece* spawnedPiece = &shapes[rand() % 7];
-        //         if(!tetris_game_is_valid_pos(tetris_state, spawnedPiece->p)) {
-        //             tetris_state->gameState = GameStateGameOver;
-        //             // processing = false;
-        //         } else {
-        //             memcpy(&tetris_state->currPiece, spawnedPiece, sizeof(tetris_state->currPiece));
-        //             osTimerStart(timer, tetris_state->fallSpeed);
-        //         }
-        //     }
-        // }
-
-        // if(tetris_game_is_valid_pos(tetris_state, newPiece->p)) {
-        //     memcpy(&tetris_state->currPiece, newPiece, sizeof(tetris_state->currPiece));
-        // }
-
-        // tetris_game_render_curr_piece(tetris_state);
 
         view_port_update(view_port);
         release_mutex(&state_mutex, tetris_state);
